requirements.txt
# Core A2A dependencies (matching HealthUniverse deployment)
a2a-sdk==0.3.0
protobuf==5.29.5

# Web server
uvicorn==0.35.0
starlette==0.47.2
fastapi==0.116.1

# LLM providers
google-adk==1.9.0
google-genai==1.28.0
langchain-google-genai==2.1.9
litellm==1.75.0
openai==1.99.1
anthropic==0.61.0

# Google Cloud dependencies
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0

# Utilities
python-dotenv==1.1.1
httpx==0.28.1
pydantic==2.11.7
structlog==25.4.0

# Other dependencies
pyyaml==6.0.2
jinja2==3.1.6
typing-extensions==4.14.1